refactor(table): extract row id and version metadata into its own module
assign_row_ids, the created_at/last_updated resolution helpers and the run
encoder all serve one concern: which row ids the new fragments get, and what
per-row version metadata travels with them. Keeping created_at correct across
an update means tracing each new row back to the fragment and offset it came
from, which is the bulk of this code and reads better away from the manifest
assembly that calls it.
